// web/src/gen/types/OpportunityResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type OpportunityResponse = {
  /**
   * @type string | undefined
   */
  color?: string
  /**
   * @type string | undefined
   */
  description?: string
  /**
   * @type string | undefined
   */
  icon?: string
  /**
   * @type string | undefined
   */
  id?: string
  /**
   * @type string | undefined
   */
  text?: string
}